rename: hook폴더의 query폴더명 및 그 안의 파일명 변경
- query폴더명을 queries로 변경
- query폴더 안의 파일명을 useComments.ts -> useCommentsQueries.ts로 변경
- query폴더 안의 파일명을 usePlaylist.ts -> usePlaylistQueries.ts로 변경
- query폴더 안의 파일명을 useYoutube.ts -> useYoutubeQueries.ts로 변경
- hooks/mutations 폴더 추가
