+++ src/pages/CommentList.tsx
@@ -10,7 +10,7 @@ import CommentBox from '@/components/comment/CommentBox';
 import IconTextButton from '@/components/common/buttons/IconTextButton';
 import Toast from '@/components/common/Toast';
 import { PATH } from '@/constants/path';
-import { useCommentsList } from '@/hooks/query/useComments';
+import { useCommentsList } from '@/hooks/queries/useCommentsQueries';
 import Header from '@/layouts/layout/Header';
 import { useToastStore } from '@/store/useToastStore';
 import theme from '@/styles/theme';
